rename: dto 및 폴더명 변경
