+++ src/main/java/cholog/supp/api/member/dto/response/MemberInfoResponse.java
@@ -2,7 +2,7 @@ package cholog.supp.api.member.dto.response;
 
 import cholog.supp.db.member.MemberType;
 
-public record MemberDataResponse(
+public record MemberInfoResponse(
     Long id,
     String email,
     MemberType memberType
