+++ src/main/java/cholog/supp/api/studygroup/dto/request/InvitationRequest.java
@@ -1,4 +1,4 @@
-package cholog.supp.api.study.dto.request;
+package cholog.supp.api.studygroup.dto.request;
 
 import cholog.supp.db.member.MemberType;
