+++ src/main/java/cholog/supp/api/studygroup/dto/response/EachGroupResponse.java
@@ -1,4 +1,4 @@
-package cholog.supp.api.study.dto.response;
+package cholog.supp.api.studygroup.dto.response;
 
 import cholog.supp.db.member.MemberType;
 import cholog.supp.db.study.StudyGroup;
